Only reevaluate compiler-related functionality in CthulhuCompilerExt
