+++ ext/CthulhuCompilerExt.jl
@@ -1,14 +1,32 @@
 module CthulhuCompilerExt
 
 using Compiler: Compiler as CC
-using Compiler.IRShow: IRShow
 using Cthulhu: Cthulhu
 
-function __init__()
-    Cthulhu.CTHULHU_MODULE[] = @__MODULE__
-    read_config!()
-end
+@static if CC.AbstractInterpreter !== Cthulhu.CC.AbstractInterpreter
+    using Compiler.IRShow: IRShow
 
-include("../src/CthulhuBase.jl")
+    using Accessors
+    using CodeTracking: CodeTracking, definition, whereis, maybe_fix_path
+    using InteractiveUtils
+    using UUIDs
+    using REPL: REPL, AbstractTerminal
+    using JuliaSyntax
+    using JuliaSyntax: SyntaxNode, AbstractSyntaxNode, child, children
+    using TypedSyntax
+    using WidthLimitedIO
+
+    using Core.IR
+
+    include("../src/CthulhuCompiler.jl")
+
+    function __init__()
+        Cthulhu.CompilerExt = @__MODULE__
+    end
+else
+    function __init__()
+        Cthulhu.CompilerExt = nothing
+    end
+end
 
 end
