+++ src/ascend.jl
@@ -1,6 +1,6 @@
 function ascend_impl(
-        term, mi;
-        interp::AbstractInterpreter=NativeInterpreter(),
+        term, mi::MethodInstance;
+        interp=Base.Compiler.NativeInterpreter(),
         provider::AbstractProvider=AbstractProvider(interp),
         pagesize::Int=10, dynamic::Bool=false, maxsize::Int=pagesize,
         menu_options=(; pagesize), kwargs...)
