+++ src/descend.jl
@@ -30,11 +30,6 @@ end
 _descend(@nospecialize(args...); terminal=default_terminal(), kwargs...) =
     _descend(terminal, args...; kwargs...)
 
-function _descend(bookmark::Bookmark; terminal=default_terminal(), kwargs...)
-    state = CthulhuState(bookmark; terminal, kwargs...)
-    descend!(state)
-end
-
 ##
 # descend! is the main driver function.
 # src/reflection.jl has the tools to discover methods.
@@ -59,7 +54,7 @@ function descend!(state::CthulhuState)
         mi::MethodInstance, ci::CodeInstance
 
         src = something(state.override, ci)
-        result = LookupResult(provider, src, config.optimize)
+        result = lookup(provider, src, config.optimize)
         if result === nothing
             if should_regenerate_code_instance(provider, src)
                 additional_descend(src)
@@ -197,10 +192,10 @@ function menu_callsites_from_source_node(callsite::Callsite, source_node)
     for info in callsite.info.callinfos
         ci = get_ci(info)
         mi = get_mi(ci)
-        p = Base.unwrap_unionall(mi.specTypes).parameters
-        if isa(source_node, TypedSyntax.MaybeTypedSyntaxNode) && length(p) == length(JuliaSyntax.children(source_node)) + 1
+        p = unwrap_unionall(mi.specTypes).parameters
+        if isa(source_node, TypedSyntax.MaybeTypedSyntaxNode) && length(p) == length(children(source_node)) + 1
             new_node = copy(source_node)
-            for (i, child) in enumerate(JuliaSyntax.children(new_node))
+            for (i, child) in enumerate(children(new_node))
                 child.typ = p[i+1]
             end
             push!(callsites, new_node)
@@ -211,7 +206,7 @@ function menu_callsites_from_source_node(callsite::Callsite, source_node)
     return callsites
 end
 
-function source_slotnames(result::LookupResult)
+function source_slotnames(result#=::LookupResult=#)
     result.src === nothing && return false
     return Base.sourceinfo_slotnames(result.src)
 end
